requirements.txt
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
backoff==2.2.1
bcrypt==5.0.0
bleach==6.3.0
bokeh==3.8.2
build==1.3.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
chromadb==1.4.0
click==8.3.1
cloudpickle==3.1.2
colorcet==3.1.0
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
dask==2025.12.0
datashader==0.18.2
decorator==5.2.1
distro==1.9.0
durationpy==0.10
executing==2.2.1
filelock==3.20.2
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2025.12.0
google-auth==2.45.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
h11==0.16.0
hf-xet==1.2.0
holoviews==1.22.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.2.3
humanfriendly==10.0
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kubernetes==34.1.0
lazy_loader==0.4
linkify-it-py==2.0.3
llvmlite==0.46.0
locket==1.0.0
lxml==6.0.2
lxml-stubs==0.5.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdit-py-plugins==0.5.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multipledispatch==1.0.0
narwhals==2.15.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
oauthlib==3.3.1
onnxruntime==1.23.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
panel==1.8.5
param==2.3.1
parso==0.8.5
partd==1.4.2
pexpect==4.9.0
pillow==12.1.0
plotly==6.5.1
posthog==5.4.0
prompt_toolkit==3.0.52
protobuf==6.33.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.3
pyct==0.6.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.1
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-frontmatter==1.1.0
pytz==2025.2
pyviz_comms==3.0.6
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.2.0
rpds-py==0.30.0
rsa==4.9.1
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.12.20
tokenizers==0.22.1
toml==0.10.2
toolz==1.1.0
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uc-micro-py==1.0.3
umap-learn==0.5.11
Unidecode==1.4.0
urllib3==2.3.0
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.14
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
xarray==2025.12.0
xyzservices==2025.11.0
zipp==3.23.0
